moved shared method to helper-class and added unit tests for it
